Create Monty
This game is based off of a game show called Let's Make a Deal and is also a probability puzzle called the Monty Hall Problem. The player is presented with three doors and asked to pick one which he/she thinks the winning prize is behind. The host will reveal what is behind one of the other two doors and ask the player if he/she would like to change their decision. This game uses the srand line to randomize the winning door.